set the version of calckey to a backup version
The "current" tag at time of writing (hash ed6ead61d7c1) is not stored
on any other tag. It is a version of 13.1.4.1 with a bug fixed that
caused performance spikes.
The repo username and password are not stored in here, as I do not
expect this to last forever.